feat(login): show login as overlay so background page stays visible
- Extract login UI into LoginOverlay.svelte controlled by loginOverlayOpen store
- Header Sign in button toggles the store instead of navigating to /login
- /login route still works for deep links (sets store on mount)
- Align login card dark-mode palette to app color tokens
- Raise login z-index above Cheffy/CookingTools widgets (9995–9997)
- Add close button to sign-in card

feat(mutes): mute list editor polish

- Move Mutable attribution to right side of toolbar before Publish button
- Reorder profile tabs: Muted before Drafts
- Expand bio show-more threshold to 400 chars / 4 lines

fix(caption): set --color-caption to blue-grey #94a3b8 in dark mode
