+++ src/lib/stores/loginOverlay.ts
@@ -1,0 +1,2 @@
+import { writable } from 'svelte/store';
+export const loginOverlayOpen = writable(false);
