+++ src/routes/+layout.svelte
@@ -15,6 +15,8 @@
   import LongformEditorModal from '../components/reads/LongformEditorModal.svelte';
   import WalletModal from '../components/wallet/WalletModal.svelte';
   import ToastContainer from '../components/ToastContainer.svelte';
+  import LoginOverlay from '../components/LoginOverlay.svelte';
+  import { loginOverlayOpen } from '$lib/stores/loginOverlay';
   import { createAuthManager, type AuthState } from '$lib/authManager';
   import { stopMessageSubscription, clearMessages } from '$lib/stores/messages';
   import { clearDecryptCache } from '$lib/encryptionService';
@@ -570,6 +572,9 @@
       <PostModal bind:open={$postComposerOpen} />
       <LongformEditorModal />
       <WalletModal />
+      {#if $loginOverlayOpen}
+        <LoginOverlay />
+      {/if}
       <ToastContainer />
     </div>
   </div>
