+++ src/routes/user/[slug]/+page.svelte
@@ -2048,12 +2048,12 @@
       <!-- Bio -->
       {#if profile?.about}
         {@const bioText = profile.about.trim()}
-        {@const needsTruncation = bioText.length > 200}
+        {@const needsTruncation = bioText.length > 400}
 
         <div class="max-w-2xl">
           <p
             class="text-sm text-caption leading-relaxed"
-            class:line-clamp-2={!bioExpanded && needsTruncation}
+            class:line-clamp-4={!bioExpanded && needsTruncation}
           >
             <ParsedBio text={bioText} />
           </p>
@@ -2192,21 +2192,6 @@
       </button>
       {#if $userPublickey && $userPublickey === hexpubkey}
         <button
-          on:click={() => (activeTab = 'drafts')}
-          class="px-4 py-2 text-sm font-medium transition-colors relative flex items-center gap-1"
-          style="color: {activeTab === 'drafts'
-            ? 'var(--color-text-primary)'
-            : 'var(--color-text-secondary)'}"
-        >
-          <FloppyDiskIcon size={16} />
-          Drafts
-          {#if activeTab === 'drafts'}
-            <span
-              class="absolute bottom-0 left-0 right-0 h-0.5 bg-gradient-to-r from-orange-500 to-amber-500"
-            ></span>
-          {/if}
-        </button>
-        <button
           on:click={() => (activeTab = 'muted')}
           class="px-4 py-2 text-sm font-medium transition-colors relative flex items-center gap-1"
           style="color: {activeTab === 'muted'
@@ -2216,6 +2201,21 @@
           <SpeakerSimpleSlashIcon size={16} />
           Muted
           {#if activeTab === 'muted'}
+            <span
+              class="absolute bottom-0 left-0 right-0 h-0.5 bg-gradient-to-r from-orange-500 to-amber-500"
+            ></span>
+          {/if}
+        </button>
+        <button
+          on:click={() => (activeTab = 'drafts')}
+          class="px-4 py-2 text-sm font-medium transition-colors relative flex items-center gap-1"
+          style="color: {activeTab === 'drafts'
+            ? 'var(--color-text-primary)'
+            : 'var(--color-text-secondary)'}"
+        >
+          <FloppyDiskIcon size={16} />
+          Drafts
+          {#if activeTab === 'drafts'}
             <span
               class="absolute bottom-0 left-0 right-0 h-0.5 bg-gradient-to-r from-orange-500 to-amber-500"
             ></span>
@@ -2523,10 +2523,10 @@
   }
 
   /* Ensure line-clamp works for bio text */
-  :global(.line-clamp-2) {
+  :global(.line-clamp-4) {
     display: -webkit-box;
-    -webkit-line-clamp: 2;
-    line-clamp: 2;
+    -webkit-line-clamp: 4;
+    line-clamp: 4;
     -webkit-box-orient: vertical;
     overflow: hidden;
   }
